+++ ui/src/pages/ConfirmEmailChange.tsx
@@ -1,5 +1,5 @@
 import { useEffect, useState } from 'react'
-import { Link } from 'react-router-dom'
+import { Link } from 'react-router'
 import { useOneTimeToken } from '../hooks/useOneTimeToken'
 import { APP_VERSION, authApi, extractApiErrorMessage } from '../api/client'
 import { BUD_LOGO_DARK, BUD_LOGO_LIGHT } from '../brandAssets'